mongobridge/requirements.txt
paho-mqtt==1.6.1
pymongo[srv]
python-dotenv==1.0.0
